+++ src/targets/browser/browserPathResolver.ts
@@ -38,6 +38,8 @@ const enum Suffix {
   Index = 'index.html',
 }
 
+const wildcardHostname = 'https?:\\/\\/[^\\/]+\\/';
+
 @injectable()
 export class BrowserSourcePathResolver extends SourcePathResolverBase<IOptions> {
   constructor(
@@ -49,20 +51,17 @@ export class BrowserSourcePathResolver extends SourcePathResolverBase<IOptions> 
     super(options, logger);
   }
 
-  absolutePathToUrl(absolutePath: string): string | undefined {
+  /** @override */
+  private absolutePathToUrlPath(absolutePath: string): { url: string; needsWildcard: boolean } {
     absolutePath = path.normalize(absolutePath);
     const { baseUrl, pathMapping } = this.options;
-    if (!baseUrl) {
-      return utils.absolutePathToFileUrl(absolutePath);
-    }
-
     const defaultMapping = ['/', pathMapping['/']] as const;
     const bestMatch =
       Object.entries(pathMapping)
         .sort(([, directoryA], [, directoryB]) => directoryB.length - directoryA.length)
         .find(([, directory]) => isSubdirectoryOf(directory, absolutePath)) || defaultMapping;
     if (!bestMatch) {
-      return utils.absolutePathToFileUrl(absolutePath);
+      return { url: utils.absolutePathToFileUrl(absolutePath), needsWildcard: false };
     }
 
     let urlPath = utils.platformPathToUrlPath(path.relative(bestMatch[1], absolutePath));
@@ -71,7 +70,11 @@ export class BrowserSourcePathResolver extends SourcePathResolverBase<IOptions> 
       urlPath = urlPrefix + '/' + urlPath;
     }
 
-    return utils.completeUrlEscapingRoot(baseUrl, urlPath);
+    if (!baseUrl && !utils.isValidUrl(urlPath)) {
+      return { url: urlPath, needsWildcard: true };
+    }
+
+    return { url: utils.completeUrlEscapingRoot(baseUrl, urlPath), needsWildcard: false };
   }
 
   public async urlToAbsolutePath({ url, map }: IUrlResolution): Promise<string | undefined> {
@@ -210,12 +213,12 @@ export class BrowserSourcePathResolver extends SourcePathResolverBase<IOptions> 
   /**
    * @override
    */
-  public absolutePathToUrlRegexp(absolutePath: string): Promise<string | undefined> {
-    let url = this.absolutePathToUrl(absolutePath);
-    if (!url) {
-      return Promise.resolve(undefined);
-    }
+  public absolutePathToUrlRegexp(absolutePath: string) {
+    const transform = this.absolutePathToUrlPath(absolutePath);
+    let url = transform.url;
 
+    // Make "index" paths optional since some servers, like vercel's serve,
+    // allow omitting them.
     let endRegexEscape = absolutePath.length;
     if (url.endsWith(Suffix.Index)) {
       endRegexEscape = url.length - Suffix.Index.length - 1;
@@ -225,6 +228,13 @@ export class BrowserSourcePathResolver extends SourcePathResolverBase<IOptions> 
       url = url.slice(0, endRegexEscape) + `(\\.html)?`;
     }
 
-    return Promise.resolve(urlToRegex(url, [0, endRegexEscape]));
+    // If there's no base URL, allow the URL to match _any_ protocol
+    let startRegexEscape = 0;
+    if (transform.needsWildcard) {
+      url = wildcardHostname + url;
+      startRegexEscape = wildcardHostname.length;
+    }
+
+    return urlToRegex(url, [startRegexEscape, endRegexEscape]);
   }
 }